commit code: Format code set alarm.
